requirements.txt
numpy
pandas
torch
scikit-learn
pyyaml
tqdm